+++ game/pieces/piece.py
@@ -1,30 +1,29 @@
-from game.board import Board
-
 class Piece:
+    
+    # Constructor de la clase
     def __init__(self, color):
-        self.__color__ = color
-        self.__board__ = Board()
+        #Constructor de la clase.
+        self._color = color
+        self._movimientos_dama_rey = [(-1, -1), (-1, 1), (1, -1), (1, 1), (-1, 0), (1, 0), (0, -1), (0, 1)]
+    
+    # Obtener el color de la pieza
+    def obtener_color(self):
+        #Esta función devuelve el valor del atributo privado _color de la pieza.
+        return self._color
 
-    @property
-    def color(self):
-        return self.__color__
-
-    def __str__(self):
-        if self.__color__ == "WHITE":
-            return self.white_str
-        else:
-            return self.black_str
-
-    @property
-    def white_str(self):
-        # Este método debe ser sobrescrito por las subclases
-        raise NotImplementedError("Subclasses should implement this method")
-
-    @property
-    def black_str(self):
-        # Este método debe ser sobrescrito por las subclases
-        raise NotImplementedError("Subclasses should implement this method")
-
-    def get_valid_moves(self, board, row, col):
-        # Este método debe ser sobrescrito por las subclases
-        raise NotImplementedError("Subclasses should implement this method")
+    # Movimientos posibles de las piezas
+    def movimientos_posibles(self, fila_inicio, columna_inicio, direcciones, un_paso=False):
+        #La función devuelve los movimientos posibles de la pieza desde su posición actual.
+        movimientos = []
+        for direccion in direcciones:
+            nueva_fila, nueva_columna = fila_inicio, columna_inicio
+            while True:
+                nueva_fila += direccion[0]
+                nueva_columna += direccion[1]
+                if 0 <= nueva_fila < 8 and 0 <= nueva_columna < 8:
+                    movimientos.append((nueva_fila, nueva_columna))
+                    if un_paso:
+                        break
+                else:
+                    break
+        return movimientos